Move load config filter to service provider
