+++ app/config/app.php
@@ -145,6 +145,7 @@ return [
         'CachetHQ\Cachet\Providers\RepositoryServiceProvider',
         'CachetHQ\Cachet\Providers\RoutingServiceProvider',
         'CachetHQ\Cachet\Providers\ViewComposerServiceProvider',
+        'CachetHQ\Cachet\Providers\LoadConfigServiceProvider',
 
     ],
